+++ docs/release-notes.md
@@ -1,6 +1,7 @@
 # Release notes
 ## 2.5
 * For players:
+  * Added install scripts for Linux/Mac (no longer need to run `mono install.exe` through the terminal).
   * When a mod is skipped because you're missing a dependency, the error now shows the name and URL of the missing mod.
   * Fixed mod crashes being logged under `[SMAPI]` instead of the mod name.
   * Fixed uninstall script not confirming success on Linux/Mac.